+++ desc/coils.py
@@ -1,6 +1,7 @@
 """Classes for magnetic field coils."""
 
 import numbers
+import os
 from abc import ABC
 from collections.abc import MutableSequence
 from functools import partial
@@ -1975,7 +1976,7 @@ class CoilSet(OptimizableCollection, _Coil, MutableSequence):
         # always start at the 3rd line after periods
         coilnames = []  # the coilgroup each coil belongs to
         # corresponds to each coil in the coilinds list
-
+        coil_file = os.path.expanduser(coil_file)
         # read in the coils file
         headind = -1
         with open(coil_file) as f:
@@ -2079,6 +2080,8 @@ class CoilSet(OptimizableCollection, _Coil, MutableSequence):
         # TODO(#1376): name each group based off of CoilSet name?
         # TODO(#1376): have CoilGroup be automatically assigned based off of
         # CoilSet if current coilset is a collection of coilsets?
+
+        coilsFilename = os.path.expanduser(coilsFilename)
 
         NFP = 1 if NFP is None else NFP